Minimax observed time

Density, distribution function, quantile function, and expected value for the minimum of the maximum observed time over two treatment arms.

dminimaxU(x, arm0, arm1, include_cens = T)

pminimaxU(q, arm0, arm1, include_cens = T, lower.tail = T)

qminimaxU(p, arm0, arm1, include_cens = T, margin = 0.01)

eminimaxU(arm0, arm1, include_cens = T)

Arguments

x, q

vector of quantiles.

arm0

object of class 'arm'.

arm1

object of class 'arm'.

include_cens

logical; if TRUE, include time-to-censoring as potential observed time; otherwise, observed time equals time-to-event.

lower.tail

logical; if TRUE, probabilities are \(P(X \le x)\); otherwise, \(P(X > x)\).

p

vector of probabilities.

margin

margin of accuracy.

Value

dminimaxU gives the density, pminimaxU gives the distribution function, qminimaxU gives the quantile function, and eminimaxU

gives the expected value.

Details

Given a patient in arm \(X_i=j\) with time-to-event \(T_i\) and time-to-censoring \(C_i\), \(U_i=\min(T_i, C_i)\) defines the patient's observed time. The maximum observed time over patients of arm \(j\) is then \(\max_{i:X_i=j} U_i\), and the minimax observed time over two arms is \(\min_j (\max_{i:X_i=j} U_i)\).

See also