Convert exponential parameters

Convert exponential survival percentile or hazard rate to the other.

per2haz(x, per = 0.5)

Arguments

x

survival percentile or exponential hazard rate

per

(per)th percentile

Value

survival percentile or hazard rate.

Details

$$y=-log(1-per)/x$$

Examples

per2haz(14)          # hazard rate for exponential with 14-month median
#> [1] 0.04951051
per2haz(0.05)        # median survival for exponential with hazard rate 0.05
#> [1] 13.86294
per2haz(14, 0.8)     # hazard rate for exponential with 80th percentile survival at 14 months
#> [1] 0.1149599
per2haz(0.27, 0.8)   # 80th percentile survival for exponential with hazard rate 0.27
#> [1] 5.960881