Power

Calculate power for a two-arm survival study.

Usage

power_two_arm(arm0, arm1, test = list(test = "weighted logrank"),
  alpha = 0.025, sides = 1)

Arguments

arm0

object of class 'arm'.

arm1

object of class 'arm'.

test

list or list of lists. Each list must contain at minimum the key 'test' describing the type of statistical test. Default test is the "weighted logrank". Kaplan-Meier based tests ("survival difference", "survival ratio", "rmst difference", "rmst ratio", "percentile difference", and "percentile ratio") require the user to define an additional key, either the desired 'milestone' or 'percentile'. The weighted log-rank test does not require additional keys. However, user may choose which weight function ("1"=unweighted, "n"=Gehan-Breslow, "sqrtN"=Tarone-Ware, "FH_p[a]_q[b]"= Fleming-Harrington with p=a and q=b) and which approximation for the large-sample mean ("asymptotic", "generalized schoenfeld", "event driven") and variance ("1", "block[ randomization]", "simple[ randomization]") they wish to use. Default choice is 'weight'="1", 'mean.approx'="asymptotic", and 'var.approx'="1". For more details regarding the different mean and variance approximations for the weight log-rank test, please see Yung and Liu (2019).

alpha

type 1 error rate

sides

1=1-sided test, 2=2-sided test

Value

power.

References

Yung, G and Liu, Y. (2019). Sample size and power for the weighted log-rank test and Kaplan-Meier based tests with allowance for non-proportional hazards. Biometrics. <doi:10.1111/biom.13196>

See also

Examples

arm0 <- create_arm(size=120, accr_time=6, surv_scale=0.05, loss_scale=0.005, follow_time=12)
arm1 <- create_arm(size=120, accr_time=6, surv_scale=0.03, loss_scale=0.005, follow_time=12)
power_two_arm(arm0, arm1)
#> [1] 0.7296235
power_two_arm(arm0, arm1, list(test="weighted logrank",
  weight="n",
  mean.approx="generalized schoenfeld",
  var.approx="block"))
#> [1] 0.7069563
power_two_arm(arm0, arm1, list(test="survival difference", milestone=12))
#> [1] 0.658848
power_two_arm(arm0, arm1, list(test="rmst ratio", milestone=12))
#> [1] 0.5684121
power_two_arm(arm0, arm1, list(test="percentile difference", percentile=0.25))
#> [1] 0.4567024
power_two_arm(arm0, arm1, list(
  list(test="weighted logrank"),
  list(test="survival difference", milestone=12)))
#>   test     power
#> 1    1 0.7296235
#> 2    2 0.6588480