Simulate complete data for a single arm

Simulate the complete data for a single arm, including time to accrual, event, and loss of follow-up. No cutoff (by number of events or time) is applied. Hence, no patients are administratively censored.

simulate_arm(arm, label = 1)

Arguments

arm

object of class 'arm'.

label

numeric label for the simulated arm, e.g. 0 for control, 1 for treatment

Value

arm

label

time.accr

time to accrual

time.obs

time to observation from accrual

time.total

time to observation from start of study

censor

0=censor, 1=event

reason

event description ('[experience ]event', '[loss to ]followup', 'administration[ censoring]')

time.surv

time to event

time.loss

time to loss of follow-up

See also

Examples

arm0 <- create_arm(size=120, accr_time=6, surv_scale=0.05, loss_scale=0.005, follow_time=12)
simulate_arm(arm0, label=0)
#>     arm  time.accr    time.obs time.total censor   reason    time.surv
#> 1     0 4.17737426  4.57869682   8.756071      1    event   4.57869682
#> 2     0 4.92005521  4.35321712   9.273272      1    event   4.35321712
#> 3     0 1.75105223 20.82056100  22.571613      1    event  20.82056100
#> 4     0 0.14071958 18.47658707  18.617307      1    event  18.47658707
#> 5     0 0.10828881 48.52242672  48.630716      1    event  48.52242672
#> 6     0 3.77495805  8.67854315  12.453501      1    event   8.67854315
#> 7     0 3.54361799  3.86509333   7.408711      1    event   3.86509333
#> 8     0 4.52354604  9.88260642  14.406152      0 followup 109.15239731
#> 9     0 2.82904103  7.91592441  10.744965      1    event   7.91592441
#> 10    0 1.13585835 28.20312887  29.338987      1    event  28.20312887
#> 11    0 2.52371997 11.65116690  14.174887      1    event  11.65116690
#> 12    0 1.56331028 10.20679526  11.770106      1    event  10.20679526
#> 13    0 1.80124517 11.62464107  13.425886      1    event  11.62464107
#> 14    0 5.06815495  0.13150928   5.199664      1    event   0.13150928
#> 15    0 4.35864685  0.34765227   4.706299      1    event   0.34765227
#> 16    0 5.57668303  6.87585787  12.452541      0 followup   7.26835592
#> 17    0 0.11832450  2.37333515   2.491660      1    event   2.37333515
#> 18    0 2.87762809  6.47380846   9.351437      1    event   6.47380846
#> 19    0 3.92034980 15.44195280  19.362303      0 followup  74.39032850
#> 20    0 4.70071119 12.91159787  17.612309      1    event  12.91159787
#> 21    0 5.66096844 29.36678971  35.027758      1    event  29.36678971
#> 22    0 2.87139955 15.75117032  18.622570      1    event  15.75117032
#> 23    0 3.51492261 80.47978509  83.994708      1    event  80.47978509
#> 24    0 0.36366019 48.28577844  48.649439      1    event  48.28577844
#> 25    0 5.12759762 11.01415915  16.141757      1    event  11.01415915
#> 26    0 0.61202145  3.26405132   3.876073      1    event   3.26405132
#> 27    0 3.22056543 20.10923982  23.329805      1    event  20.10923982
#> 28    0 3.23982702 14.62324885  17.863076      1    event  14.62324885
#> 29    0 0.71048981 10.27357670  10.984067      1    event  10.27357670
#> 30    0 4.43270473 25.09878474  29.531489      0 followup  45.08211080
#> 31    0 0.93159697 11.11235997  12.043957      1    event  11.11235997
#> 32    0 0.96633898  0.32312957   1.289469      1    event   0.32312957
#> 33    0 0.07583187 20.38763697  20.463469      1    event  20.38763697
#> 34    0 5.96870258  3.13676688   9.105469      1    event   3.13676688
#> 35    0 0.18786230  2.89443010   3.082292      1    event   2.89443010
#> 36    0 2.12078040  2.02013653   4.140917      0 followup   2.28877604
#> 37    0 0.53528427 31.92214577  32.457430      1    event  31.92214577
#> 38    0 3.87191488  4.75574811   8.627663      1    event   4.75574811
#> 39    0 4.16625711 22.35065924  26.516916      1    event  22.35065924
#> 40    0 1.33031590 14.32425212  15.654568      1    event  14.32425212
#> 41    0 2.07651432 74.05352751  76.130042      1    event  74.05352751
#> 42    0 4.71742863  1.84660715   6.564036      0 followup  23.26698706
#> 43    0 3.11433218  2.72292962   5.837262      1    event   2.72292962
#> 44    0 1.91616011  3.34576175   5.261922      1    event   3.34576175
#> 45    0 1.49980438 18.01084272  19.510647      1    event  18.01084272
#> 46    0 0.23137440  6.99032854   7.221703      0 followup  15.13491397
#> 47    0 3.94786788  6.10990929  10.057777      1    event   6.10990929
#> 48    0 5.79186404 16.73686684  22.528731      1    event  16.73686684
#> 49    0 0.62620782 40.73354058  41.359748      1    event  40.73354058
#> 50    0 1.22854575 23.11437878  24.342925      1    event  23.11437878
#> 51    0 3.93572416 42.29378550  46.229510      1    event  42.29378550
#> 52    0 1.56551939  1.16155573   2.727075      1    event   1.16155573
#> 53    0 5.60309653  0.06049528   5.663592      1    event   0.06049528
#> 54    0 3.00183964 13.14708867  16.148928      1    event  13.14708867
#> 55    0 1.96497481 22.43670750  24.401682      1    event  22.43670750
#> 56    0 5.93160503  5.09903358  11.030639      1    event   5.09903358
#> 57    0 2.22581489 24.44086926  26.666684      1    event  24.44086926
#> 58    0 2.04476170 12.61101986  14.655782      1    event  12.61101986
#> 59    0 3.08797758 16.47280004  19.560778      1    event  16.47280004
#> 60    0 2.65087456 42.05248284  44.703357      1    event  42.05248284
#> 61    0 3.68602328  1.25370561   4.939729      1    event   1.25370561
#> 62    0 5.71883755 17.18472972  22.903567      1    event  17.18472972
#> 63    0 1.34039492  3.84628564   5.186681      0 followup  37.35799223
#> 64    0 5.44184541 59.36287462  64.804720      0 followup  71.21875146
#> 65    0 4.46351515 12.29981080  16.763326      1    event  12.29981080
#> 66    0 4.11071698 31.10027292  35.210990      1    event  31.10027292
#> 67    0 1.58188250  8.43208832  10.013971      1    event   8.43208832
#> 68    0 4.76976004 25.06925011  29.839010      1    event  25.06925011
#> 69    0 0.40793235  1.13319197   1.541124      1    event   1.13319197
#> 70    0 5.55640150  0.66395611   6.220358      1    event   0.66395611
#> 71    0 2.70205980 21.65290805  24.354968      1    event  21.65290805
#> 72    0 4.92478938 12.80138974  17.726179      1    event  12.80138974
#> 73    0 2.24936662 33.70539700  35.954764      1    event  33.70539700
#> 74    0 0.74474274 26.68992544  27.434668      1    event  26.68992544
#> 75    0 2.13497557  6.30581971   8.440795      1    event   6.30581971
#> 76    0 2.23359852 47.54649800  49.780097      1    event  47.54649800
#> 77    0 2.81434104  5.82381145   8.638152      1    event   5.82381145
#> 78    0 3.87990905 47.00949885  50.889408      1    event  47.00949885
#> 79    0 3.90004271 28.59907198  32.499115      1    event  28.59907198
#> 80    0 5.61734913 11.33324892  16.950598      1    event  11.33324892
#> 81    0 3.26944223 55.52410029  58.793543      1    event  55.52410029
#> 82    0 1.46313481  0.55294470   2.016080      1    event   0.55294470
#> 83    0 5.55362793  1.56801026   7.121638      1    event   1.56801026
#> 84    0 0.33934182  6.82602393   7.165366      1    event   6.82602393
#> 85    0 0.40907990 20.45506116  20.864141      1    event  20.45506116
#> 86    0 4.68465033 47.98817473  52.672825      1    event  47.98817473
#> 87    0 4.32246845 47.28987329  51.612342      0 followup  84.91362793
#> 88    0 5.88770784  2.73042582   8.618134      1    event   2.73042582
#> 89    0 5.91530434 21.02201096  26.937315      1    event  21.02201096
#> 90    0 0.86008985 45.94590554  46.805995      1    event  45.94590554
#> 91    0 5.62051987 16.61681231  22.237332      1    event  16.61681231
#> 92    0 3.20900268  4.03980758   7.248810      0 followup  36.08709726
#> 93    0 4.90774048 11.60748838  16.515229      0 followup  21.15497956
#> 94    0 4.68207735 19.84245118  24.524529      1    event  19.84245118
#> 95    0 2.86267797 18.23804868  21.100727      1    event  18.23804868
#> 96    0 5.82220803 18.73384834  24.556056      1    event  18.73384834
#> 97    0 4.23581466  5.96688568  10.202700      0 followup  25.69971832
#> 98    0 4.95083845 16.74867541  21.699514      1    event  16.74867541
#> 99    0 3.08493992  2.82127068   5.906211      1    event   2.82127068
#> 100   0 0.33880199  3.21736908   3.556171      0 followup  84.10580857
#> 101   0 4.20672440  0.35578541   4.562510      1    event   0.35578541
#> 102   0 0.78420260 15.26601029  16.050213      1    event  15.26601029
#> 103   0 0.73204559  3.28418478   4.016230      1    event   3.28418478
#> 104   0 5.27263151  5.76081358  11.033445      1    event   5.76081358
#> 105   0 4.22942749 36.57291542  40.802343      1    event  36.57291542
#> 106   0 1.02297664 52.91143453  53.934411      1    event  52.91143453
#> 107   0 2.88335706  7.58348028  10.466837      1    event   7.58348028
#> 108   0 2.44075245 24.37847996  26.819232      1    event  24.37847996
#> 109   0 1.74649599 63.88486775  65.631364      1    event  63.88486775
#> 110   0 5.21954672 22.31108287  27.530630      1    event  22.31108287
#> 111   0 1.55640383  0.33786304   1.894267      1    event   0.33786304
#> 112   0 2.00255145  7.60723385   9.609785      0 followup  30.24916138
#> 113   0 4.81207831  7.83653522  12.648614      1    event   7.83653522
#> 114   0 3.89765910  2.19541296   6.093072      1    event   2.19541296
#> 115   0 4.94059579 15.12616740  20.066763      1    event  15.12616740
#> 116   0 5.80664790 15.21169310  21.018341      0 followup  15.31516305
#> 117   0 4.12970960  0.36129789   4.491007      1    event   0.36129789
#> 118   0 1.01791492 53.12709928  54.145014      0 followup  55.41087966
#> 119   0 2.44300870 68.15309937  70.596108      1    event  68.15309937
#> 120   0 1.62272100 10.51508244  12.137803      0 followup  78.47105689
#>      time.loss
#> 1   188.775019
#> 2   100.812669
#> 3   522.602116
#> 4   122.204786
#> 5    96.083178
#> 6   118.723771
#> 7    28.557412
#> 8     9.882606
#> 9   374.873267
#> 10  290.879624
#> 11  397.619874
#> 12   87.944692
#> 13   44.697172
#> 14   33.359090
#> 15  189.737705
#> 16    6.875858
#> 17   27.808886
#> 18  154.141453
#> 19   15.441953
#> 20   82.405572
#> 21  260.886372
#> 22   65.280604
#> 23  114.395209
#> 24  263.814396
#> 25   93.256851
#> 26   68.868027
#> 27  222.985256
#> 28  856.160932
#> 29  147.127594
#> 30   25.098785
#> 31   17.612025
#> 32  194.333074
#> 33   64.672034
#> 34   93.286277
#> 35  347.153975
#> 36    2.020137
#> 37  266.714791
#> 38   72.430187
#> 39   74.459422
#> 40  138.411828
#> 41  386.210366
#> 42    1.846607
#> 43  111.114403
#> 44  186.926103
#> 45   77.085750
#> 46    6.990329
#> 47   42.503777
#> 48  643.703910
#> 49   48.051117
#> 50  391.689498
#> 51  181.292441
#> 52  316.162180
#> 53   79.426394
#> 54   30.610879
#> 55  112.583875
#> 56  276.331541
#> 57  169.795711
#> 58  203.100993
#> 59  226.738543
#> 60  336.926494
#> 61  177.374308
#> 62   92.958596
#> 63    3.846286
#> 64   59.362875
#> 65  232.077408
#> 66  192.768407
#> 67  411.559407
#> 68  301.572566
#> 69    2.511514
#> 70  474.879228
#> 71  196.438502
#> 72  517.044171
#> 73   66.957352
#> 74   69.392545
#> 75  558.720930
#> 76   93.502902
#> 77   33.651157
#> 78  309.002342
#> 79   72.843861
#> 80  136.007479
#> 81  180.740942
#> 82   17.348399
#> 83   41.579616
#> 84   50.666789
#> 85  398.418970
#> 86  154.198103
#> 87   47.289873
#> 88   72.545017
#> 89  305.401750
#> 90  352.980554
#> 91  118.375228
#> 92    4.039808
#> 93   11.607488
#> 94  152.439620
#> 95  301.531005
#> 96  999.547763
#> 97    5.966886
#> 98   40.599602
#> 99  266.609707
#> 100   3.217369
#> 101  66.453895
#> 102 300.661590
#> 103 231.491738
#> 104 217.573611
#> 105  90.124963
#> 106  67.521417
#> 107 282.589290
#> 108 139.255889
#> 109 324.556725
#> 110 256.835193
#> 111 206.612282
#> 112   7.607234
#> 113 142.862685
#> 114 124.629928
#> 115  94.012542
#> 116  15.211693
#> 117  44.897537
#> 118  53.127099
#> 119 136.631645
#> 120  10.515082