Simulate a clinical trial

Simulate a single- or two-arm clinical trial, where end of study (EOS) is triggered after a number of events has been observed or a certain time has elapsed. Whereas simulate_arm provides complete data for patients, including time to event and loss of follow-up, simulate_trial mimicks an actual survival study by providing only the observed time (minimum of time to event or censoring) and censoring indicator.

simulate_trial(arm0 = NA, arm1 = NA, events = NA, duration = Inf)

Arguments

arm0

object of class 'arm'.

arm1

object of class 'arm'.

events

number of required events to trigger end of study; overrides study duration defined within arm0 and arm1.

duration

time from first-patient-in to trigger end of study; overrides study duration defined within arm0 and arm1. If both events and duration are specified, end of study is triggered by either criteria, whichever occurs first.

Value

arm

0=arm0, 1=arm1

time.accr

time to accrual

time.obs

time to observation from accrual

time.total

time to observation from start of study

censor

0=censor, 1=event

reason

event description ('[experience ]event', '[loss to ]followup', 'administration[ censoring]')

See also

Examples

arm0 <- create_arm(size=120, accr_time=6, surv_scale=0.05, loss_scale=0.005, follow_time=12)
arm1 <- create_arm(size=120, accr_time=6, surv_scale=0.03, loss_scale=0.005, follow_time=12)
simulate_trial(arm0, duration=10)
#>     arm  time.accr    time.obs time.total censor         reason
#> 93    0 0.22599342 1.653212949   1.879206      1          event
#> 109   0 0.27092737 1.832236499   2.103164      1          event
#> 8     0 1.55959925 0.610449261   2.170049      0       followup
#> 44    0 2.43293496 0.003882628   2.436818      0       followup
#> 74    0 0.07178863 2.557585526   2.629374      1          event
#> 89    0 0.53167770 2.309488006   2.841166      1          event
#> 59    0 1.43484733 2.223401830   3.658249      1          event
#> 119   0 3.29189938 0.665745355   3.957645      0       followup
#> 57    0 1.79931262 2.202794528   4.002107      0       followup
#> 115   0 1.65446074 2.361306358   4.015767      1          event
#> 27    0 3.97560129 0.572020836   4.547622      1          event
#> 58    0 1.84135188 2.807610842   4.648963      1          event
#> 66    0 4.98240996 0.046830339   5.029240      1          event
#> 51    0 1.21595581 3.815888027   5.031844      1          event
#> 43    0 4.21049710 1.393216312   5.603713      1          event
#> 45    0 1.87314205 3.900244652   5.773387      1          event
#> 1     0 4.03734058 1.779598757   5.816939      1          event
#> 67    0 3.04740784 2.790048579   5.837456      1          event
#> 29    0 3.83803437 2.143068579   5.981103      1          event
#> 90    0 3.13028885 3.071636797   6.201926      1          event
#> 114   0 3.92740296 2.274894263   6.202297      1          event
#> 37    0 0.07960416 6.572219894   6.651824      1          event
#> 111   0 1.34629129 5.372373235   6.718665      1          event
#> 96    0 3.90287380 2.941647698   6.844522      1          event
#> 101   0 3.74757630 3.154108132   6.901684      1          event
#> 25    0 4.78761776 2.525362196   7.312980      1          event
#> 77    0 3.18802182 4.299788593   7.487810      1          event
#> 68    0 3.07984646 4.448576410   7.528423      1          event
#> 24    0 2.86547476 5.019952342   7.885427      1          event
#> 118   0 1.58678969 6.306998897   7.893789      1          event
#> 106   0 2.42992846 5.496984546   7.926913      0       followup
#> 112   0 4.88912365 3.068792570   7.957916      1          event
#> 12    0 4.00402902 4.117689099   8.121718      1          event
#> 5     0 1.23003512 6.923003320   8.153038      1          event
#> 19    0 4.40878312 3.808128791   8.216912      1          event
#> 6     0 4.08332506 4.136049778   8.219375      0       followup
#> 83    0 4.56569482 4.279930245   8.845625      1          event
#> 42    0 5.53373291 3.429635702   8.963369      0       followup
#> 53    0 4.59926844 4.618505392   9.217774      1          event
#> 40    0 4.19674222 5.315711187   9.512453      1          event
#> 92    0 3.70558717 5.966120573   9.671708      1          event
#> 103   0 2.17872546 7.502913580   9.681639      1          event
#> 116   0 2.17991127 7.820088728  10.000000      0 administration
#> 95    0 1.19941416 8.800585835  10.000000      0 administration
#> 60    0 3.76575655 6.234243453  10.000000      0 administration
#> 99    0 0.08886243 9.911137573  10.000000      0 administration
#> 39    0 0.77650295 9.223497052  10.000000      0 administration
#> 21    0 5.81219214 4.187807856  10.000000      0 administration
#> 31    0 5.27181769 4.728182310  10.000000      0 administration
#> 120   0 4.76120864 5.238791362  10.000000      0 administration
#> 55    0 3.50389903 6.496100967  10.000000      0 administration
#> 78    0 5.47280064 4.527199363  10.000000      0 administration
#> 14    0 2.36710683 7.632893170  10.000000      0 administration
#> 87    0 5.35852305 4.641476952  10.000000      0 administration
#> 100   0 4.60989458 5.390105421  10.000000      0 administration
#> 91    0 0.49975360 9.500246398  10.000000      0 administration
#> 105   0 1.24030814 8.759691861  10.000000      0 administration
#> 18    0 3.15186120 6.848138796  10.000000      0 administration
#> 26    0 1.87642004 8.123579958  10.000000      0 administration
#> 35    0 4.72000033 5.279999669  10.000000      0 administration
#> 7     0 4.02679052 5.973209478  10.000000      0 administration
#> 76    0 5.38646469 4.613535310  10.000000      0 administration
#> 108   0 1.65995790 8.340042103  10.000000      0 administration
#> 104   0 5.96732611 4.032673885  10.000000      0 administration
#> 107   0 3.86461227 6.135387727  10.000000      0 administration
#> 13    0 5.14330592 4.856694075  10.000000      0 administration
#> 11    0 5.55027929 4.449720706  10.000000      0 administration
#> 113   0 5.84085113 4.159148875  10.000000      0 administration
#> 72    0 0.18151542 9.818484577  10.000000      0 administration
#> 15    0 1.13877875 8.861221249  10.000000      0 administration
#> 33    0 1.57338322 8.426616779  10.000000      0 administration
#> 69    0 5.61366962 4.386330382  10.000000      0 administration
#> 41    0 4.79851365 5.201486347  10.000000      0 administration
#> 36    0 3.41787555 6.582124452  10.000000      0 administration
#> 30    0 3.30789501 6.692104994  10.000000      0 administration
#> 88    0 4.99015653 5.009843471  10.000000      0 administration
#> 34    0 3.11972270 6.880277297  10.000000      0 administration
#> 3     0 1.55272700 8.447273002  10.000000      0 administration
#> 75    0 4.66127070 5.338729301  10.000000      0 administration
#> 85    0 1.66350282 8.336497185  10.000000      0 administration
#> 50    0 3.94653573 6.053464265  10.000000      0 administration
#> 94    0 0.59438019 9.405619811  10.000000      0 administration
#> 73    0 3.48608942 6.513910583  10.000000      0 administration
#> 79    0 2.47479740 7.525202598  10.000000      0 administration
#> 97    0 4.11097350 5.889026500  10.000000      0 administration
#> 4     0 5.48190419 4.518095810  10.000000      0 administration
#> 84    0 5.08293509 4.917064907  10.000000      0 administration
#> 65    0 3.65343555 6.346564453  10.000000      0 administration
#> 80    0 4.61954080 5.380459202  10.000000      0 administration
#> 70    0 3.35995471 6.640045286  10.000000      0 administration
#> 71    0 4.23702155 5.762978454  10.000000      0 administration
#> 54    0 4.40961888 5.590381125  10.000000      0 administration
#> 9     0 0.43020539 9.569794608  10.000000      0 administration
#> 17    0 0.32912596 9.670874043  10.000000      0 administration
#> 63    0 5.70176499 4.298235006  10.000000      0 administration
#> 38    0 1.21058337 8.789416625  10.000000      0 administration
#> 81    0 5.71863327 4.281366735  10.000000      0 administration
#> 28    0 0.97254591 9.027454092  10.000000      0 administration
#> 22    0 4.09778848 5.902211519  10.000000      0 administration
#> 82    0 0.71033019 9.289669811  10.000000      0 administration
#> 61    0 4.74736778 5.252632222  10.000000      0 administration
#> 16    0 5.13135332 4.868646683  10.000000      0 administration
#> 32    0 5.30987097 4.690129031  10.000000      0 administration
#> 48    0 2.98787228 7.012127720  10.000000      0 administration
#> 102   0 5.77591578 4.224084220  10.000000      0 administration
#> 23    0 4.50750135 5.492498649  10.000000      0 administration
#> 2     0 1.84422445 8.155775550  10.000000      0 administration
#> 98    0 5.71419309 4.285806911  10.000000      0 administration
#> 10    0 5.89021895 4.109781046  10.000000      0 administration
#> 56    0 1.30528292 8.694717081  10.000000      0 administration
#> 47    0 4.60560303 5.394396969  10.000000      0 administration
#> 46    0 0.99620688 9.003793124  10.000000      0 administration
#> 64    0 3.80939830 6.190601696  10.000000      0 administration
#> 20    0 1.08780024 8.912199757  10.000000      0 administration
#> 52    0 5.58016608 4.419833920  10.000000      0 administration
#> 62    0 3.08972750 6.910272505  10.000000      0 administration
#> 49    0 2.37632917 7.623670828  10.000000      0 administration
#> 110   0 3.69873560 6.301264399  10.000000      0 administration
#> 117   0 0.86422985 9.135770145  10.000000      0 administration
#> 86    0 2.79780776 7.202192240  10.000000      0 administration
simulate_trial(arm0, arm1, events=50)
#>     arm  time.accr  time.obs time.total censor         reason
#> 111   0 0.02562488 1.3167098   1.342335      1          event
#> 43    0 0.28761876 1.1601792   1.447798      1          event
#> 121   1 0.48017591 1.0017516   1.481927      1          event
#> 210   1 1.16997536 0.5772301   1.747205      1          event
#> 240   1 1.49147131 0.8197120   2.311183      1          event
#> 77    0 0.87646143 1.5330509   2.409512      1          event
#> 36    0 2.28164855 0.2882547   2.569903      1          event
#> 205   1 0.62881825 2.0602971   2.689115      1          event
#> 54    0 2.03413751 0.7333969   2.767534      1          event
#> 88    0 0.89926761 1.9146745   2.813942      1          event
#> 47    0 2.69986264 0.2359026   2.935765      1          event
#> 113   0 0.46690507 2.7861474   3.253053      1          event
#> 155   1 1.19878941 2.2394302   3.438220      1          event
#> 165   1 1.73471227 1.7105629   3.445275      1          event
#> 167   1 3.21121469 0.2542042   3.465419      1          event
#> 46    0 2.47012900 1.1894978   3.659627      1          event
#> 126   1 0.91322459 2.8123750   3.725600      1          event
#> 64    0 3.81355384 0.1887540   4.002308      1          event
#> 19    0 1.63348767 2.4702017   4.103689      1          event
#> 45    0 2.45433796 1.8106735   4.265011      1          event
#> 32    0 2.64934567 1.8249422   4.474288      1          event
#> 39    0 3.92936227 0.6063606   4.535723      1          event
#> 2     0 2.52093741 2.1899574   4.710895      1          event
#> 99    0 4.44260007 0.4016478   4.844248      1          event
#> 27    0 2.28258646 2.5976989   4.880285      1          event
#> 207   1 0.69149311 4.2093860   4.900879      1          event
#> 204   1 2.32319575 2.7882978   5.111494      1          event
#> 115   0 3.37988145 1.9374465   5.317328      1          event
#> 217   1 3.97881758 1.3431775   5.321995      0       followup
#> 30    0 2.75188965 2.6474536   5.399343      1          event
#> 69    0 3.16392606 2.3426648   5.506591      1          event
#> 108   0 1.24121261 4.3192907   5.560503      1          event
#> 56    0 2.53860450 3.0414579   5.580062      1          event
#> 133   1 2.32885979 3.2622704   5.591130      1          event
#> 38    0 4.20518216 1.4877906   5.692973      1          event
#> 67    0 1.32279300 4.3716294   5.694422      1          event
#> 20    0 3.59379653 2.1327448   5.726541      0       followup
#> 97    0 5.06055781 0.7500303   5.810588      1          event
#> 123   1 5.15338005 0.8071192   5.960499      1          event
#> 216   1 1.80749993 4.2624094   6.069909      0       followup
#> 148   1 3.65528469 2.4165962   6.071881      1          event
#> 117   0 0.10254963 6.0001946   6.102744      1          event
#> 224   1 0.46698764 6.1308379   6.597826      1          event
#> 190   1 1.69332079 5.0764736   6.769794      1          event
#> 25    0 1.93112807 4.9791987   6.910327      1          event
#> 127   1 4.66015989 2.4860157   7.146176      1          event
#> 118   0 3.91549116 3.2938220   7.209313      0       followup
#> 201   1 2.59877576 4.6742238   7.273000      1          event
#> 221   1 5.65346350 1.6739405   7.327404      1          event
#> 186   1 1.67649275 5.6757639   7.352257      1          event
#> 59    0 4.81898903 2.7885464   7.607535      1          event
#> 142   1 5.65579666 1.9644749   7.620272      1          event
#> 209   1 0.89752238 6.9731542   7.870677      1          event
#> 208   1 1.54963691 6.3563668   7.906004      1          event
#> 235   1 0.67648677 7.2295169   7.906004      0 administration
#> 72    0 5.45380849 2.4521952   7.906004      0 administration
#> 94    0 3.91160632 3.9943974   7.906004      0 administration
#> 92    0 3.05326463 4.8527391   7.906004      0 administration
#> 22    0 4.74264579 3.1633579   7.906004      0 administration
#> 229   1 1.84768468 6.0583190   7.906004      0 administration
#> 98    0 4.39392935 3.5120743   7.906004      0 administration
#> 40    0 2.25655677 5.6494469   7.906004      0 administration
#> 140   1 2.89421423 5.0117895   7.906004      0 administration
#> 4     0 3.71756045 4.1884432   7.906004      0 administration
#> 219   1 2.31395911 5.5920446   7.906004      0 administration
#> 68    0 5.73295371 2.1730500   7.906004      0 administration
#> 189   1 2.83789186 5.0681118   7.906004      0 administration
#> 159   1 3.53169076 4.3743129   7.906004      0 administration
#> 83    0 0.46830211 7.4377016   7.906004      0 administration
#> 37    0 1.12310949 6.7828942   7.906004      0 administration
#> 175   1 2.17815580 5.7278479   7.906004      0 administration
#> 93    0 1.41640079 6.4896029   7.906004      0 administration
#> 166   1 0.61337642 7.2926273   7.906004      0 administration
#> 223   1 2.43110638 5.4748973   7.906004      0 administration
#> 1     0 1.53119721 6.3748065   7.906004      0 administration
#> 132   1 1.00757421 6.8984295   7.906004      0 administration
#> 103   0 0.02333992 7.8826638   7.906004      0 administration
#> 3     0 5.03886584 2.8671379   7.906004      0 administration
#> 145   1 3.06855772 4.8374460   7.906004      0 administration
#> 157   1 2.24143496 5.6645687   7.906004      0 administration
#> 214   1 2.46903625 5.4369674   7.906004      0 administration
#> 185   1 3.46646340 4.4395403   7.906004      0 administration
#> 135   1 4.79515886 3.1108448   7.906004      0 administration
#> 194   1 0.75333427 7.1526694   7.906004      0 administration
#> 70    0 5.62710208 2.2789016   7.906004      0 administration
#> 211   1 3.65399897 4.2520047   7.906004      0 administration
#> 82    0 5.82367662 2.0823271   7.906004      0 administration
#> 61    0 5.55968401 2.3463197   7.906004      0 administration
#> 65    0 5.82334953 2.0826542   7.906004      0 administration
#> 44    0 3.81312614 4.0928776   7.906004      0 administration
#> 21    0 5.51355554 2.3924481   7.906004      0 administration
#> 178   1 2.02938331 5.8766204   7.906004      0 administration
#> 182   1 5.97271304 1.9332906   7.906004      0 administration
#> 73    0 2.68747758 5.2185261   7.906004      0 administration
#> 179   1 5.03623966 2.8697640   7.906004      0 administration
#> 96    0 4.93433220 2.9716715   7.906004      0 administration
#> 164   1 4.06437287 3.8416308   7.906004      0 administration
#> 84    0 3.77727923 4.1287245   7.906004      0 administration
#> 42    0 3.97094858 3.9350551   7.906004      0 administration
#> 151   1 4.85202656 3.0539771   7.906004      0 administration
#> 193   1 2.32636560 5.5796381   7.906004      0 administration
#> 81    0 4.79497672 3.1110270   7.906004      0 administration
#> 12    0 0.59886381 7.3071399   7.906004      0 administration
#> 114   0 0.46949425 7.4365094   7.906004      0 administration
#> 62    0 1.37812687 6.5278768   7.906004      0 administration
#> 149   1 2.77292100 5.1330827   7.906004      0 administration
#> 14    0 5.77356839 2.1324353   7.906004      0 administration
#> 169   1 0.45097588 7.4550278   7.906004      0 administration
#> 239   1 4.07397590 3.8320278   7.906004      0 administration
#> 196   1 0.46956078 7.4364429   7.906004      0 administration
#> 10    0 5.88133357 2.0246701   7.906004      0 administration
#> 5     0 2.94521778 4.9607859   7.906004      0 administration
#> 6     0 5.95419494 1.9518088   7.906004      0 administration
#> 192   1 3.30301578 4.6029879   7.906004      0 administration
#> 48    0 5.47142222 2.4345815   7.906004      0 administration
#> 87    0 2.40663093 5.4993728   7.906004      0 administration
#> 35    0 2.65186241 5.2541413   7.906004      0 administration
#> 131   1 5.81589593 2.0901078   7.906004      0 administration
#> 79    0 0.57799335 7.3280103   7.906004      0 administration
#> 163   1 5.86162994 2.0443737   7.906004      0 administration
#> 18    0 0.32832291 7.5776808   7.906004      0 administration
#> 233   1 4.50478273 3.4012210   7.906004      0 administration
#> 116   0 2.42895835 5.4770453   7.906004      0 administration
#> 86    0 4.37607949 3.5299242   7.906004      0 administration
#> 120   0 4.17450758 3.7314961   7.906004      0 administration
#> 184   1 2.37967457 5.5263291   7.906004      0 administration
#> 137   1 0.89165254 7.0143512   7.906004      0 administration
#> 63    0 0.43773994 7.4682638   7.906004      0 administration
#> 195   1 3.10300112 4.8030026   7.906004      0 administration
#> 130   1 2.92169785 4.9843058   7.906004      0 administration
#> 13    0 5.07527723 2.8307265   7.906004      0 administration
#> 162   1 3.06371555 4.8422881   7.906004      0 administration
#> 215   1 5.98589985 1.9201038   7.906004      0 administration
#> 218   1 2.60995734 5.2960464   7.906004      0 administration
#> 89    0 3.65156890 4.2544348   7.906004      0 administration
#> 16    0 4.34239216 3.5636115   7.906004      0 administration
#> 29    0 5.86170149 2.0443022   7.906004      0 administration
#> 51    0 2.88497306 5.0210306   7.906004      0 administration
#> 91    0 2.02499033 5.8810134   7.906004      0 administration
#> 80    0 1.93326459 5.9727391   7.906004      0 administration
#> 226   1 3.89324109 4.0127626   7.906004      0 administration
#> 172   1 3.71577130 4.1902324   7.906004      0 administration
#> 74    0 4.48910838 3.4168953   7.906004      0 administration
#> 34    0 2.55236670 5.3536370   7.906004      0 administration
#> 76    0 0.87783031 7.0281734   7.906004      0 administration
#> 109   0 2.52641546 5.3795882   7.906004      0 administration
#> 238   1 5.19900677 2.7069969   7.906004      0 administration
#> 28    0 3.60519384 4.3008099   7.906004      0 administration
#> 170   1 5.30141844 2.6045853   7.906004      0 administration
#> 191   1 1.87527570 6.0307280   7.906004      0 administration
#> 128   1 0.61233246 7.2936712   7.906004      0 administration
#> 24    0 4.22708533 3.6789184   7.906004      0 administration
#> 71    0 4.71318366 3.1928200   7.906004      0 administration
#> 85    0 3.69046704 4.2155367   7.906004      0 administration
#> 26    0 5.73494144 2.1710623   7.906004      0 administration
#> 66    0 1.17742455 6.7285791   7.906004      0 administration
#> 225   1 0.38419804 7.5218057   7.906004      0 administration
#> 7     0 5.19661259 2.7093911   7.906004      0 administration
#> 57    0 4.89765059 3.0083531   7.906004      0 administration
#> 9     0 2.78361091 5.1223928   7.906004      0 administration
#> 31    0 1.07072377 6.8352799   7.906004      0 administration
#> 227   1 2.82036069 5.0856430   7.906004      0 administration
#> 23    0 4.78612055 3.1198831   7.906004      0 administration
#> 156   1 1.73819628 6.1678074   7.906004      0 administration
#> 199   1 4.77240889 3.1335948   7.906004      0 administration
#> 160   1 3.03382948 4.8721742   7.906004      0 administration
#> 143   1 0.01817582 7.8878279   7.906004      0 administration
#> 146   1 5.34803408 2.5579696   7.906004      0 administration
#> 17    0 1.61199288 6.2940108   7.906004      0 administration
#> 231   1 1.47322991 6.4327738   7.906004      0 administration
#> 154   1 4.66491274 3.2410910   7.906004      0 administration
#> 187   1 2.86265277 5.0433509   7.906004      0 administration
#> 181   1 3.90110807 4.0048956   7.906004      0 administration
#> 153   1 5.62059263 2.2854111   7.906004      0 administration
#> 58    0 2.10324432 5.8027594   7.906004      0 administration
#> 55    0 3.60213299 4.3038707   7.906004      0 administration
#> 8     0 3.85804867 4.0479550   7.906004      0 administration
#> 230   1 1.86334042 6.0426633   7.906004      0 administration
#> 105   0 0.45688041 7.4491233   7.906004      0 administration
#> 188   1 5.06995852 2.8360452   7.906004      0 administration
#> 41    0 1.46174280 6.4442609   7.906004      0 administration
#> 107   0 3.05146994 4.8545338   7.906004      0 administration
#> 122   1 1.87687721 6.0291265   7.906004      0 administration
#> 53    0 5.03311148 2.8728922   7.906004      0 administration
#> 236   1 4.25749835 3.6485053   7.906004      0 administration
#> 11    0 5.11758627 2.7884174   7.906004      0 administration
#> 112   0 2.75552713 5.1504766   7.906004      0 administration
#> 141   1 3.29124536 4.6147583   7.906004      0 administration
#> 232   1 2.44337051 5.4626332   7.906004      0 administration
#> 106   0 5.70232918 2.2036745   7.906004      0 administration
#> 183   1 5.39175622 2.5142475   7.906004      0 administration
#> 90    0 0.61141231 7.2945914   7.906004      0 administration
#> 60    0 3.38457119 4.5214325   7.906004      0 administration
#> 75    0 5.97330060 1.9327031   7.906004      0 administration
#> 213   1 5.92019184 1.9858119   7.906004      0 administration
#> 228   1 3.00658505 4.8994186   7.906004      0 administration
#> 33    0 3.81616875 4.0898349   7.906004      0 administration
#> 102   0 1.36092937 6.5450743   7.906004      0 administration
#> 161   1 3.25199986 4.6540038   7.906004      0 administration
#> 100   0 0.35312881 7.5528749   7.906004      0 administration
#> 101   0 2.31388232 5.5921214   7.906004      0 administration
#> 50    0 5.45180431 2.4541994   7.906004      0 administration
#> 222   1 2.24150324 5.6645005   7.906004      0 administration
#> 136   1 2.20908465 5.6969190   7.906004      0 administration
#> 78    0 2.37851170 5.5274920   7.906004      0 administration
#> 49    0 5.70135765 2.2046460   7.906004      0 administration
#> 134   1 1.28542551 6.6205782   7.906004      0 administration
#> 177   1 5.63831838 2.2676853   7.906004      0 administration
#> 197   1 4.80810149 3.0979022   7.906004      0 administration
#> 95    0 2.36160784 5.5443959   7.906004      0 administration
#> 147   1 0.93934617 6.9666575   7.906004      0 administration
#> 168   1 1.51248249 6.3935212   7.906004      0 administration
#> 198   1 0.04413513 7.8618686   7.906004      0 administration
#> 220   1 2.79905521 5.1069485   7.906004      0 administration
#> 200   1 1.80748953 6.0985142   7.906004      0 administration
#> 206   1 4.49510188 3.4109018   7.906004      0 administration
#> 125   1 3.83736717 4.0686365   7.906004      0 administration
#> 180   1 3.36808276 4.5379209   7.906004      0 administration
#> 139   1 1.56230733 6.3436964   7.906004      0 administration
#> 150   1 3.39060719 4.5153965   7.906004      0 administration
#> 119   0 5.57462047 2.3313832   7.906004      0 administration
#> 52    0 3.10729666 4.7987070   7.906004      0 administration
#> 212   1 5.12983234 2.7761714   7.906004      0 administration
#> 124   1 1.24573711 6.6602666   7.906004      0 administration
#> 104   0 3.10206758 4.8039361   7.906004      0 administration
#> 176   1 5.16669317 2.7393105   7.906004      0 administration
#> 234   1 5.43793825 2.4680654   7.906004      0 administration
#> 203   1 5.91823387 1.9877698   7.906004      0 administration
#> 237   1 5.64955420 2.2564495   7.906004      0 administration
#> 110   0 2.71818781 5.1878159   7.906004      0 administration
#> 15    0 2.83950768 5.0664960   7.906004      0 administration
#> 138   1 4.98957771 2.9164260   7.906004      0 administration
#> 158   1 2.49882493 5.4071788   7.906004      0 administration
#> 152   1 2.11834986 5.7876538   7.906004      0 administration
#> 173   1 2.74333507 5.1626686   7.906004      0 administration
#> 144   1 1.67955467 6.2264490   7.906004      0 administration
#> 129   1 1.42604893 6.4799548   7.906004      0 administration
#> 202   1 3.83725294 4.0687507   7.906004      0 administration
#> 171   1 1.47666209 6.4293416   7.906004      0 administration
#> 174   1 2.70008050 5.2059232   7.906004      0 administration