Create a 'lachin' object

Create an object of class 'lachin' by specifying the trial parameters for a single arm, including the sample size, accrual distribution, survival distribution, loss to follow-up distribution, and study duration. 'Lachin' objects are also 'arm' objects, but with accrual limited to the uniform and truncated exponential distributions, and survival and loss to follow-up limited to the exponential distribution. 'Lachin' objects have the advantage that expectations for certain counting processes have closed form equations and can therefore be calculated more efficiently (Lachin and Foulkes, 1986).

Usage

create_arm_lachin(size, accr_time, accr_dist = "pieceuni",
  accr_param = NA, surv_median = NA, surv_exphazard = NA,
  surv_milestone = NA, loss_median = NA, loss_exphazard = NA,
  loss_milestone = NA, follow_time = Inf, total_time = Inf)

Arguments

size

sample size. If total sample size is unknown, provide the integer sample size relative to the opposing arm, e.g. 1 for 1:2 randomization ratio or 2 for 2:3.

accr_time

accrual duration.

accr_dist

accrual distribution. Default is uniform (piecewise uniform with one interval). Alternatively, 'truncexp' allows for a truncated exponential distribution as proposed by Lachin and Foulkes (1986). Depending on the value of accr_param, this distribution can be either convex or concave.

accr_param

additional accrual parameter for accr_dist='truncexp'. accr_param>0 specifies a convex distribution and accr_param<0 specifies a concave distribution.

surv_median

median survival.

surv_exphazard

exponential hazard rate for the survival distribution.

surv_milestone

a tuple c(milestone, probability) that uniquely defines the exponential survival distribution, e.g. c(12, 0.8) corresponds to the exponential distribution with 80% survival rate at 12 months.

loss_median

median loss to follow-up.

loss_exphazard

exponential hazard rate for the loss to follow-up distribution.

loss_milestone

a tuple c(milestone, probability) that uniquely defines the exponential loss to follow-up distribution, e.g. c(12, 0.99) corresponds to the exponential distribution with 1% loss to follow-up at 12 months.

follow_time

Follow-up duration. Either follow_time or total_time (below) should be specified.

total_time

Total study duration. Either follow_time (above) or total_time should be specified.

Value

a list containing assumptions of size, accrual, censoring, survival, and follow-up for a single arm.

References

Lachin, J. M. and Foulkes, M. A. (1986) Evaluation of sample size and power for analyses of survival with allowance for nonuniform patient entry, losses to follow-up, noncompliance, and stratification. Biometrics, 42, 507-519.

See also

Examples

# 3 arms with similar survival and loss to follow-up
create_arm_lachin(size=120, accr_time=6,
  surv_median=14,
  loss_median=140,
  follow_time=12)
#> $size
#> [1] 120
#> 
#> $accr_time
#> [1] 6
#> 
#> $accr_dist
#> [1] "pieceuni"
#> 
#> $accr_interval
#> [1] 0 6
#> 
#> $accr_param
#> [1] NA
#> 
#> $surv_cure
#> [1] 0
#> 
#> $surv_interval
#> [1]   0 Inf
#> 
#> $surv_shape
#> [1] 1
#> 
#> $surv_scale
#> [1] 0.04951051
#> 
#> $loss_shape
#> [1] 1
#> 
#> $loss_scale
#> [1] 0.004951051
#> 
#> $follow_time
#> [1] 12
#> 
#> $total_time
#> [1] 18
#> 
#> attr(,"class")
#> [1] "list"   "lachin" "arm"   
create_arm_lachin(size=120, accr_time=6,
  surv_exphazard=0.05,
  loss_exphazard=0.005,
  follow_time=12)
#> $size
#> [1] 120
#> 
#> $accr_time
#> [1] 6
#> 
#> $accr_dist
#> [1] "pieceuni"
#> 
#> $accr_interval
#> [1] 0 6
#> 
#> $accr_param
#> [1] NA
#> 
#> $surv_cure
#> [1] 0
#> 
#> $surv_interval
#> [1]   0 Inf
#> 
#> $surv_shape
#> [1] 1
#> 
#> $surv_scale
#> [1] 0.05
#> 
#> $loss_shape
#> [1] 1
#> 
#> $loss_scale
#> [1] 0.005
#> 
#> $follow_time
#> [1] 12
#> 
#> $total_time
#> [1] 18
#> 
#> attr(,"class")
#> [1] "list"   "lachin" "arm"   
create_arm_lachin(size=120, accr_time=6,
  accr_dist="truncexp",
  accr_param=0.1,
  surv_milestone=c(14, 0.5),
  loss_milestone=c(140, 0.5),
  total_time=18)
#> $size
#> [1] 120
#> 
#> $accr_time
#> [1] 6
#> 
#> $accr_dist
#> [1] "truncexp"
#> 
#> $accr_interval
#> [1] 0 6
#> 
#> $accr_param
#> [1] 0.1
#> 
#> $surv_cure
#> [1] 0
#> 
#> $surv_interval
#> [1]   0 Inf
#> 
#> $surv_shape
#> [1] 1
#> 
#> $surv_scale
#> [1] 0.04951051
#> 
#> $loss_shape
#> [1] 1
#> 
#> $loss_scale
#> [1] 0.004951051
#> 
#> $follow_time
#> [1] 12
#> 
#> $total_time
#> [1] 18
#> 
#> attr(,"class")
#> [1] "list"   "lachin" "arm"