Simulate complete data for a single arm

Simulate the complete data for a single arm, including time to accrual, event, and loss of follow-up. No cutoff (by number of events or time) is applied. Hence, no patients are administratively censored.

Usage

simulate_arm(arm, label = 1)

Arguments

arm

object of class 'arm'.

label

numeric label for the simulated arm, e.g. 0 for control, 1 for treatment

Value

arm

label

time.accr

time to accrual

time.obs

time to observation from accrual

time.total

time to observation from start of study

censor

0=censor, 1=event

reason

event description ('[experience ]event', '[loss to ]followup', 'administration[ censoring]')

time.surv

time to event

time.loss

time to loss of follow-up

See also

Examples

arm0 <- create_arm(size=120, accr_time=6, surv_scale=0.05, loss_scale=0.005, follow_time=12)
simulate_arm(arm0, label=0)
#>     arm  time.accr    time.obs time.total censor   reason   time.surv
#> 1     0 3.94310877  0.98845284   4.931562      1    event  0.98845284
#> 2     0 2.03327735  9.38424610  11.417523      1    event  9.38424610
#> 3     0 2.50158976 34.01727889  36.518869      1    event 34.01727889
#> 4     0 1.28831833 29.85911313  31.147431      0 followup 54.51116770
#> 5     0 3.56348960 11.03236658  14.595856      1    event 11.03236658
#> 6     0 2.93101725  2.58804071   5.519058      1    event  2.58804071
#> 7     0 0.39755765 38.00281683  38.400374      1    event 38.00281683
#> 8     0 5.67001549 15.28146116  20.951477      1    event 15.28146116
#> 9     0 3.08461143 29.23446896  32.319080      1    event 29.23446896
#> 10    0 5.50060878  6.98025025  12.480859      1    event  6.98025025
#> 11    0 5.74399860 55.47659666  61.220595      0 followup 68.86026214
#> 12    0 5.66489220 46.06026160  51.725154      1    event 46.06026160
#> 13    0 3.99937284 17.24971604  21.249089      1    event 17.24971604
#> 14    0 3.01850363  9.56314981  12.581653      1    event  9.56314981
#> 15    0 0.43511717 17.19021812  17.625335      1    event 17.19021812
#> 16    0 2.39248986 12.87459771  15.267088      1    event 12.87459771
#> 17    0 2.73160160 38.18452661  40.916128      1    event 38.18452661
#> 18    0 4.35289828 49.50651526  53.859414      1    event 49.50651526
#> 19    0 1.52353033 16.79000404  18.313534      1    event 16.79000404
#> 20    0 2.13497381 47.49303662  49.628010      1    event 47.49303662
#> 21    0 2.93915166 22.77371472  25.712866      1    event 22.77371472
#> 22    0 5.96368819  2.44969089   8.413379      1    event  2.44969089
#> 23    0 0.65610490 36.15007401  36.806179      1    event 36.15007401
#> 24    0 1.98520913 15.78802164  17.773231      1    event 15.78802164
#> 25    0 3.82807892 90.82311757  94.651196      1    event 90.82311757
#> 26    0 4.01917982 62.61922958  66.638409      0 followup 74.36471688
#> 27    0 4.68898668 15.71447678  20.403463      1    event 15.71447678
#> 28    0 2.79705989  4.25672962   7.053790      0 followup  8.12030247
#> 29    0 0.17164938 12.66424120  12.835891      1    event 12.66424120
#> 30    0 0.80541641  4.55241911   5.357836      1    event  4.55241911
#> 31    0 2.88460728  1.79102706   4.675634      1    event  1.79102706
#> 32    0 2.64726864 39.15206789  41.799337      1    event 39.15206789
#> 33    0 5.22422827  9.52106408  14.745292      1    event  9.52106408
#> 34    0 2.04443764 17.45731458  19.501752      1    event 17.45731458
#> 35    0 2.01727286  0.30268299   2.319956      1    event  0.30268299
#> 36    0 0.43798136  0.96598586   1.403967      0 followup 43.42097396
#> 37    0 3.28890539 35.15138541  38.440291      1    event 35.15138541
#> 38    0 3.15472322 22.45453443  25.609258      1    event 22.45453443
#> 39    0 5.09557794  8.25674182  13.352320      1    event  8.25674182
#> 40    0 2.65653047 42.59659712  45.253128      1    event 42.59659712
#> 41    0 2.58376371 13.24656592  15.830330      1    event 13.24656592
#> 42    0 2.26744630  3.22933988   5.496786      0 followup  9.85681823
#> 43    0 5.83315862 16.25880086  22.091959      1    event 16.25880086
#> 44    0 2.72062292 72.37572204  75.096345      1    event 72.37572204
#> 45    0 2.35820095 45.03129669  47.389498      0 followup 59.63735842
#> 46    0 2.79449653  0.40109786   3.195594      1    event  0.40109786
#> 47    0 4.10677688  0.08822255   4.194999      1    event  0.08822255
#> 48    0 3.79229363  6.91137048  10.703664      0 followup  9.93215656
#> 49    0 1.77644326 18.54340210  20.319845      1    event 18.54340210
#> 50    0 4.33105171  2.49855047   6.829602      1    event  2.49855047
#> 51    0 4.45231127 18.63191893  23.084230      1    event 18.63191893
#> 52    0 3.16786382  1.38911104   4.556975      1    event  1.38911104
#> 53    0 5.55941128  8.67728199  14.236693      1    event  8.67728199
#> 54    0 3.27687916 22.66310495  25.939984      1    event 22.66310495
#> 55    0 0.45163151  0.99214958   1.443781      1    event  0.99214958
#> 56    0 3.28213486 35.13663912  38.418774      1    event 35.13663912
#> 57    0 0.55368080 16.12320073  16.676882      0 followup 20.61624680
#> 58    0 4.97253462 19.51154689  24.484082      1    event 19.51154689
#> 59    0 5.48385413 15.05629185  20.540146      1    event 15.05629185
#> 60    0 0.88559602 10.28255580  11.168152      0 followup 15.29314405
#> 61    0 3.54660805 54.82590856  58.372517      1    event 54.82590856
#> 62    0 5.49534020 19.16163376  24.656974      1    event 19.16163376
#> 63    0 3.90026665 34.00678439  37.907051      1    event 34.00678439
#> 64    0 0.94285204  9.88648639  10.829338      1    event  9.88648639
#> 65    0 2.97445044  8.99237146  11.966822      1    event  8.99237146
#> 66    0 2.78694594 37.05856455  39.845510      1    event 37.05856455
#> 67    0 4.14313783  0.41725581   4.560394      1    event  0.41725581
#> 68    0 0.88885462  7.10372898   7.992584      1    event  7.10372898
#> 69    0 0.04508444  1.20853305   1.253617      1    event  1.20853305
#> 70    0 4.40289918 19.88184399  24.284743      1    event 19.88184399
#> 71    0 5.66806836 61.24152418  66.909593      1    event 61.24152418
#> 72    0 3.37460456 19.57253736  22.947142      1    event 19.57253736
#> 73    0 0.05040628  3.64764473   3.698051      1    event  3.64764473
#> 74    0 0.96180266  6.78473163   7.746534      1    event  6.78473163
#> 75    0 1.76031070 10.73366814  12.493979      1    event 10.73366814
#> 76    0 3.06054112 15.87699034  18.937531      1    event 15.87699034
#> 77    0 1.90860090 25.49027457  27.398875      1    event 25.49027457
#> 78    0 0.20639885 55.85485334  56.061252      1    event 55.85485334
#> 79    0 2.66897229 19.38802978  22.057002      1    event 19.38802978
#> 80    0 2.80306993 11.58451079  14.387581      1    event 11.58451079
#> 81    0 4.86566007 32.67261244  37.538273      1    event 32.67261244
#> 82    0 1.87577938 48.38732121  50.263101      1    event 48.38732121
#> 83    0 3.93283149 18.03004637  21.962878      1    event 18.03004637
#> 84    0 0.26121565  4.29051228   4.551728      1    event  4.29051228
#> 85    0 4.69818919  4.07775615   8.775945      1    event  4.07775615
#> 86    0 5.64296475 13.55657924  19.199544      1    event 13.55657924
#> 87    0 0.22013050 17.48199331  17.702124      1    event 17.48199331
#> 88    0 5.96910517 23.10018448  29.069290      1    event 23.10018448
#> 89    0 4.96256808  0.50564292   5.468211      1    event  0.50564292
#> 90    0 3.63381359  2.65951229   6.293326      0 followup 20.04324336
#> 91    0 1.71055983  2.07977586   3.790336      1    event  2.07977586
#> 92    0 2.47621177 22.02549599  24.501708      1    event 22.02549599
#> 93    0 3.25807410 64.60843160  67.866506      1    event 64.60843160
#> 94    0 3.57730955  8.19009905  11.767409      1    event  8.19009905
#> 95    0 3.37973852  3.04004110   6.419780      1    event  3.04004110
#> 96    0 3.87490002 41.64301759  45.517918      1    event 41.64301759
#> 97    0 0.07833281  1.52966049   1.607993      1    event  1.52966049
#> 98    0 3.74940017  7.80983765  11.559238      1    event  7.80983765
#> 99    0 3.25518289 38.62810948  41.883292      1    event 38.62810948
#> 100   0 4.95870086 45.07060201  50.029303      1    event 45.07060201
#> 101   0 5.71977822  8.61279680  14.332575      1    event  8.61279680
#> 102   0 0.35996114 22.36409767  22.724059      1    event 22.36409767
#> 103   0 2.75672838 15.34228326  18.099012      1    event 15.34228326
#> 104   0 4.92110677 21.90669172  26.827798      1    event 21.90669172
#> 105   0 0.32822299 15.46298947  15.791212      1    event 15.46298947
#> 106   0 3.08403651  4.05000527   7.134042      1    event  4.05000527
#> 107   0 4.17468632 22.01620664  26.190893      1    event 22.01620664
#> 108   0 0.99023756 19.18688536  20.177123      0 followup 44.31957517
#> 109   0 0.45966377  3.29379271   3.753456      1    event  3.29379271
#> 110   0 5.42178421  4.14760342   9.569388      1    event  4.14760342
#> 111   0 5.14900420 10.80325645  15.952261      1    event 10.80325645
#> 112   0 5.61194156 15.32494068  20.936882      1    event 15.32494068
#> 113   0 3.36234585 34.86598831  38.228334      0 followup 53.07746756
#> 114   0 2.25529571  7.36335783   9.618654      1    event  7.36335783
#> 115   0 1.23586805  2.67835417   3.914222      1    event  2.67835417
#> 116   0 0.88460494 21.33473535  22.219340      1    event 21.33473535
#> 117   0 4.06913261 30.97029318  35.039426      0 followup 32.37862507
#> 118   0 4.97996967 38.66923433  43.649204      1    event 38.66923433
#> 119   0 1.73910192  1.39572545   3.134827      1    event  1.39572545
#> 120   0 1.70140380 30.14889952  31.850303      1    event 30.14889952
#>        time.loss
#> 1    195.8853600
#> 2     73.2213756
#> 3    436.9451217
#> 4     29.8591131
#> 5     15.6779871
#> 6     25.6240482
#> 7     52.2984642
#> 8     60.1847994
#> 9     44.7045582
#> 10   539.8744330
#> 11    55.4765967
#> 12    87.9916962
#> 13   452.2108471
#> 14   205.2699688
#> 15    90.7050572
#> 16   331.5137640
#> 17   147.1101706
#> 18   732.8209922
#> 19    25.6582469
#> 20   100.9707026
#> 21   285.8033935
#> 22   404.1576669
#> 23   169.9627426
#> 24   194.6718329
#> 25   580.0199675
#> 26    62.6192296
#> 27    19.0549929
#> 28     4.2567296
#> 29   228.3755121
#> 30   411.7209680
#> 31   240.4921566
#> 32   114.7949343
#> 33   197.8650894
#> 34   104.7176140
#> 35   241.4231596
#> 36     0.9659859
#> 37  1255.6196424
#> 38   199.7064847
#> 39   517.6081238
#> 40    62.7577975
#> 41   142.8236451
#> 42     3.2293399
#> 43   543.5174579
#> 44    80.7040361
#> 45    45.0312967
#> 46    36.9900816
#> 47   154.5674061
#> 48     6.9113705
#> 49    53.8187737
#> 50   194.2038060
#> 51   276.8522885
#> 52   287.1265592
#> 53    10.2214103
#> 54   154.6309102
#> 55   466.8263383
#> 56    61.0229625
#> 57    16.1232007
#> 58   566.5415518
#> 59   299.3602316
#> 60    10.2825558
#> 61   124.6847678
#> 62   591.9546876
#> 63    34.1466739
#> 64   126.9501387
#> 65    80.0074706
#> 66    55.5682644
#> 67   291.0275746
#> 68     7.2125018
#> 69    65.1360365
#> 70    41.1909336
#> 71   117.4624143
#> 72   148.7380541
#> 73    18.9417491
#> 74   124.2567739
#> 75   188.9480321
#> 76   820.8785856
#> 77    72.3685007
#> 78   532.3703920
#> 79   263.0334108
#> 80   102.3583277
#> 81    39.7137039
#> 82    58.7202159
#> 83    74.6453665
#> 84    18.6130706
#> 85    36.1834749
#> 86  1050.6961181
#> 87   170.1694746
#> 88    35.0592400
#> 89    32.9857665
#> 90     2.6595123
#> 91    65.3464899
#> 92   285.6313004
#> 93   192.5569994
#> 94    71.1282976
#> 95   188.4184292
#> 96   385.6667926
#> 97   215.8209612
#> 98   208.7581229
#> 99   417.0520188
#> 100  344.9592232
#> 101   32.0466179
#> 102   77.2871648
#> 103  322.4889067
#> 104  301.7831199
#> 105  227.9438408
#> 106   85.8628759
#> 107  302.7266622
#> 108   19.1868854
#> 109   78.3088285
#> 110   46.6367226
#> 111  145.5850196
#> 112  227.5528637
#> 113   34.8659883
#> 114  201.2151893
#> 115  454.6808671
#> 116  442.6130464
#> 117   30.9702932
#> 118  150.8983264
#> 119   22.8320615
#> 120  110.3606654