Simulate a clinical trial

Simulate a single- or two-arm clinical trial, where end of study (EOS) is triggered after a number of events has been observed or a certain time has elapsed. Whereas simulate_arm provides complete data for patients, including time to event and loss of follow-up, simulate_trial mimicks an actual survival study by providing only the observed time (minimum of time to event or censoring) and censoring indicator.

Usage

simulate_trial(arm0 = NA, arm1 = NA, events = NA, duration = Inf)

Arguments

arm0

object of class 'arm'.

arm1

object of class 'arm'.

events

number of required events to trigger end of study; overrides study duration defined within arm0 and arm1.

duration

time from first-patient-in to trigger end of study; overrides study duration defined within arm0 and arm1. If both events and duration are specified, end of study is triggered by either criteria, whichever occurs first.

Value

arm

0=arm0, 1=arm1

time.accr

time to accrual

time.obs

time to observation from accrual

time.total

time to observation from start of study

censor

0=censor, 1=event

reason

event description ('[experience ]event', '[loss to ]followup', 'administration[ censoring]')

See also

Examples

arm0 <- create_arm(size=120, accr_time=6, surv_scale=0.05, loss_scale=0.005, follow_time=12)
arm1 <- create_arm(size=120, accr_time=6, surv_scale=0.03, loss_scale=0.005, follow_time=12)
simulate_trial(arm0, duration=10)
#>     arm   time.accr   time.obs time.total censor         reason
#> 113   0 0.338500697 0.02714828  0.3656490      1          event
#> 103   0 0.499259889 0.15808126  0.6573411      1          event
#> 84    0 0.356565418 0.61243220  0.9689976      1          event
#> 12    0 0.182220431 1.90257023  2.0847907      1          event
#> 88    0 2.648214980 0.24557556  2.8937905      1          event
#> 75    0 1.697325689 1.43129951  3.1286252      1          event
#> 74    0 3.200544338 0.15368721  3.3542316      0       followup
#> 33    0 3.625704073 0.51110929  4.1368134      1          event
#> 13    0 2.541602022 1.66134332  4.2029453      1          event
#> 97    0 0.233417527 4.62685789  4.8602754      1          event
#> 91    0 1.786052138 3.23869853  5.0247507      1          event
#> 52    0 4.655595365 0.49668530  5.1522807      1          event
#> 67    0 0.191812829 5.00182746  5.1936403      1          event
#> 81    0 4.430762898 2.01204464  6.4428075      0       followup
#> 104   0 3.073110921 3.94130564  7.0144166      1          event
#> 70    0 3.286676100 3.88760279  7.1742789      1          event
#> 120   0 4.738342349 2.46470359  7.2030459      1          event
#> 117   0 5.568248022 1.69140066  7.2596487      1          event
#> 38    0 2.398359869 4.92858087  7.3269407      1          event
#> 51    0 0.204506564 7.86193063  8.0664372      1          event
#> 7     0 4.811798891 3.32244184  8.1342407      1          event
#> 24    0 1.760207820 6.57164697  8.3318548      1          event
#> 65    0 2.652614243 5.87415045  8.5267647      1          event
#> 23    0 1.801330718 6.93152309  8.7328538      1          event
#> 55    0 1.663539919 7.74380732  9.4073472      1          event
#> 11    0 3.455046231 5.98586646  9.4409127      1          event
#> 18    0 1.572655804 8.13404732  9.7067031      1          event
#> 26    0 3.799957527 6.04667401  9.8466315      1          event
#> 78    0 0.369899949 9.63010005 10.0000000      0 administration
#> 100   0 2.521216532 7.47878347 10.0000000      0 administration
#> 106   0 2.598160524 7.40183948 10.0000000      0 administration
#> 31    0 0.924036670 9.07596333 10.0000000      0 administration
#> 6     0 1.367811483 8.63218852 10.0000000      0 administration
#> 57    0 3.455050767 6.54494923 10.0000000      0 administration
#> 110   0 2.170172618 7.82982738 10.0000000      0 administration
#> 61    0 3.739133761 6.26086624 10.0000000      0 administration
#> 41    0 0.045745765 9.95425424 10.0000000      0 administration
#> 34    0 3.117232413 6.88276759 10.0000000      0 administration
#> 99    0 1.384032486 8.61596751 10.0000000      0 administration
#> 98    0 5.880262922 4.11973708 10.0000000      0 administration
#> 71    0 3.700719452 6.29928055 10.0000000      0 administration
#> 30    0 3.385457165 6.61454284 10.0000000      0 administration
#> 92    0 3.016673137 6.98332686 10.0000000      0 administration
#> 73    0 4.253430823 5.74656918 10.0000000      0 administration
#> 10    0 2.396772866 7.60322713 10.0000000      0 administration
#> 82    0 4.628988070 5.37101193 10.0000000      0 administration
#> 119   0 4.568359905 5.43164010 10.0000000      0 administration
#> 53    0 0.980275844 9.01972416 10.0000000      0 administration
#> 17    0 2.136403411 7.86359659 10.0000000      0 administration
#> 62    0 0.156190622 9.84380938 10.0000000      0 administration
#> 85    0 5.104807387 4.89519261 10.0000000      0 administration
#> 93    0 0.807781046 9.19221895 10.0000000      0 administration
#> 44    0 5.431861382 4.56813862 10.0000000      0 administration
#> 64    0 2.774639157 7.22536084 10.0000000      0 administration
#> 3     0 2.292626832 7.70737317 10.0000000      0 administration
#> 83    0 2.698721352 7.30127865 10.0000000      0 administration
#> 72    0 4.011363122 5.98863688 10.0000000      0 administration
#> 45    0 3.446518052 6.55348195 10.0000000      0 administration
#> 59    0 0.129384795 9.87061520 10.0000000      0 administration
#> 58    0 2.036152059 7.96384794 10.0000000      0 administration
#> 68    0 1.092718947 8.90728105 10.0000000      0 administration
#> 5     0 1.042302231 8.95769777 10.0000000      0 administration
#> 96    0 4.335615129 5.66438487 10.0000000      0 administration
#> 49    0 4.522233026 5.47776697 10.0000000      0 administration
#> 63    0 2.018209254 7.98179075 10.0000000      0 administration
#> 28    0 5.228397748 4.77160225 10.0000000      0 administration
#> 48    0 5.546600887 4.45339911 10.0000000      0 administration
#> 22    0 1.290033395 8.70996660 10.0000000      0 administration
#> 114   0 3.754192298 6.24580770 10.0000000      0 administration
#> 29    0 2.447484564 7.55251544 10.0000000      0 administration
#> 37    0 0.005208163 9.99479184 10.0000000      0 administration
#> 77    0 3.286963791 6.71303621 10.0000000      0 administration
#> 76    0 4.979228036 5.02077196 10.0000000      0 administration
#> 1     0 0.940184503 9.05981550 10.0000000      0 administration
#> 8     0 1.440828006 8.55917199 10.0000000      0 administration
#> 14    0 2.340922118 7.65907788 10.0000000      0 administration
#> 69    0 2.098606656 7.90139334 10.0000000      0 administration
#> 32    0 0.464178545 9.53582146 10.0000000      0 administration
#> 89    0 0.456461638 9.54353836 10.0000000      0 administration
#> 60    0 5.697093856 4.30290614 10.0000000      0 administration
#> 25    0 5.983637167 4.01636283 10.0000000      0 administration
#> 108   0 1.329179910 8.67082009 10.0000000      0 administration
#> 47    0 3.062572789 6.93742721 10.0000000      0 administration
#> 115   0 2.124613521 7.87538648 10.0000000      0 administration
#> 42    0 1.946482960 8.05351704 10.0000000      0 administration
#> 4     0 2.099828109 7.90017189 10.0000000      0 administration
#> 66    0 3.386920736 6.61307926 10.0000000      0 administration
#> 95    0 3.038939090 6.96106091 10.0000000      0 administration
#> 118   0 4.262585641 5.73741436 10.0000000      0 administration
#> 36    0 4.648208231 5.35179177 10.0000000      0 administration
#> 39    0 4.005747900 5.99425210 10.0000000      0 administration
#> 80    0 1.831329355 8.16867064 10.0000000      0 administration
#> 79    0 3.341274855 6.65872514 10.0000000      0 administration
#> 43    0 4.695760846 5.30423915 10.0000000      0 administration
#> 21    0 4.714453380 5.28554662 10.0000000      0 administration
#> 111   0 3.285927342 6.71407266 10.0000000      0 administration
#> 86    0 0.870924017 9.12907598 10.0000000      0 administration
#> 101   0 2.883892866 7.11610713 10.0000000      0 administration
#> 9     0 1.984875621 8.01512438 10.0000000      0 administration
#> 46    0 3.961810534 6.03818947 10.0000000      0 administration
#> 107   0 3.798061252 6.20193875 10.0000000      0 administration
#> 102   0 2.314174385 7.68582562 10.0000000      0 administration
#> 109   0 3.221792657 6.77820734 10.0000000      0 administration
#> 2     0 2.549878470 7.45012153 10.0000000      0 administration
#> 35    0 2.310122247 7.68987775 10.0000000      0 administration
#> 54    0 4.049908339 5.95009166 10.0000000      0 administration
#> 90    0 3.649046587 6.35095341 10.0000000      0 administration
#> 16    0 3.177506019 6.82249398 10.0000000      0 administration
#> 19    0 0.625978142 9.37402186 10.0000000      0 administration
#> 87    0 0.134992833 9.86500717 10.0000000      0 administration
#> 112   0 0.896036376 9.10396362 10.0000000      0 administration
#> 116   0 2.050919825 7.94908017 10.0000000      0 administration
#> 27    0 0.497384618 9.50261538 10.0000000      0 administration
#> 50    0 5.322180970 4.67781903 10.0000000      0 administration
#> 94    0 2.842728883 7.15727112 10.0000000      0 administration
#> 15    0 5.331545105 4.66845489 10.0000000      0 administration
#> 105   0 0.049393520 9.95060648 10.0000000      0 administration
#> 20    0 5.360596870 4.63940313 10.0000000      0 administration
#> 56    0 2.061893803 7.93810620 10.0000000      0 administration
#> 40    0 1.259315209 8.74068479 10.0000000      0 administration
simulate_trial(arm0, arm1, events=50)
#>     arm  time.accr   time.obs time.total censor         reason
#> 55    0 0.97259767 1.57809420   2.550692      1          event
#> 199   1 0.28055407 2.69373741   2.974291      0       followup
#> 63    0 2.81674180 0.16869917   2.985441      1          event
#> 79    0 0.70106747 2.80713664   3.508204      1          event
#> 173   1 3.42911867 0.09094229   3.520061      1          event
#> 171   1 3.02967281 0.72427791   3.753951      1          event
#> 175   1 1.52245955 2.26019879   3.782658      1          event
#> 9     0 3.22333552 0.59265798   3.815994      1          event
#> 90    0 3.20547095 1.02919842   4.234669      1          event
#> 184   1 0.23713378 4.22425555   4.461389      0       followup
#> 31    0 0.07051216 4.64399176   4.714504      1          event
#> 81    0 3.15889638 1.59755802   4.756454      1          event
#> 114   0 3.97254550 0.83906706   4.811613      1          event
#> 136   1 3.97114941 1.13887763   5.110027      1          event
#> 135   1 0.88451476 4.35295523   5.237470      1          event
#> 213   1 0.98984174 4.25518155   5.245023      1          event
#> 72    0 3.74037581 1.56366525   5.304041      1          event
#> 218   1 2.97874879 2.51950618   5.498255      0       followup
#> 48    0 1.16476365 4.40034990   5.565114      1          event
#> 54    0 3.28652245 2.32722924   5.613752      1          event
#> 50    0 0.33714164 5.27700987   5.614152      1          event
#> 224   1 4.06419061 2.01047395   6.074665      1          event
#> 169   1 1.75924182 4.38512575   6.144368      1          event
#> 129   1 1.15414832 5.01696783   6.171116      0       followup
#> 11    0 2.64862198 3.66092752   6.309550      1          event
#> 67    0 2.33748677 3.99964140   6.337128      1          event
#> 240   1 3.74663394 2.69099695   6.437631      1          event
#> 89    0 3.89887012 2.67622524   6.575095      1          event
#> 37    0 5.38423128 1.23356997   6.617801      1          event
#> 41    0 1.61197296 5.01092829   6.622901      0       followup
#> 239   1 0.59808121 6.12676569   6.724847      0       followup
#> 233   1 2.59746254 4.30296741   6.900430      1          event
#> 45    0 5.97092827 1.10197734   7.072906      1          event
#> 165   1 0.65522309 6.47009662   7.125320      1          event
#> 113   0 4.98660986 2.27721461   7.263824      1          event
#> 118   0 5.87928228 1.65554852   7.534831      1          event
#> 139   1 5.32796442 2.27645925   7.604424      1          event
#> 180   1 5.38799950 2.25124253   7.639242      1          event
#> 57    0 2.35367843 5.40073796   7.754416      1          event
#> 4     0 0.75504846 7.00364777   7.758696      1          event
#> 93    0 4.69559076 3.22294256   7.918533      1          event
#> 97    0 1.94853385 6.08060964   8.029143      1          event
#> 69    0 3.00767506 5.09522503   8.102900      1          event
#> 194   1 2.94288774 5.20531709   8.148205      1          event
#> 102   0 3.44219840 4.71940663   8.161605      1          event
#> 207   1 2.29234040 5.95762018   8.249961      1          event
#> 22    0 5.67177607 2.64974519   8.321521      1          event
#> 56    0 4.08825317 4.27190525   8.360158      1          event
#> 222   1 2.96036338 5.40298426   8.363348      1          event
#> 62    0 4.83017483 3.83527648   8.665451      0       followup
#> 209   1 3.80115130 5.07894313   8.880094      1          event
#> 6     0 4.83394967 4.09737405   8.931324      1          event
#> 141   1 1.66235388 7.35772953   9.020083      0       followup
#> 151   1 4.22400902 4.92784628   9.151855      1          event
#> 42    0 0.51168915 8.80627479   9.317964      1          event
#> 101   0 3.41921901 5.94982277   9.369042      1          event
#> 73    0 3.65285332 6.02555778   9.678411      0       followup
#> 15    0 0.33781302 9.40099384   9.738807      1          event
#> 127   1 4.04137046 5.76451650   9.805887      1          event
#> 163   1 5.81833658 3.98755038   9.805887      0 administration
#> 76    0 2.25117117 7.55471579   9.805887      0 administration
#> 13    0 3.98619908 5.81968788   9.805887      0 administration
#> 83    0 4.89378120 4.91210575   9.805887      0 administration
#> 188   1 3.75169984 6.05418712   9.805887      0 administration
#> 77    0 0.67028715 9.13559981   9.805887      0 administration
#> 235   1 3.93244442 5.87344254   9.805887      0 administration
#> 228   1 3.75662435 6.04926260   9.805887      0 administration
#> 38    0 0.99195132 8.81393564   9.805887      0 administration
#> 236   1 5.73275077 4.07313618   9.805887      0 administration
#> 65    0 3.32742503 6.47846193   9.805887      0 administration
#> 146   1 3.61652536 6.18936159   9.805887      0 administration
#> 3     0 5.82906418 3.97682278   9.805887      0 administration
#> 193   1 3.60855466 6.19733230   9.805887      0 administration
#> 86    0 1.91536733 7.89051963   9.805887      0 administration
#> 219   1 2.10003547 7.70585148   9.805887      0 administration
#> 112   0 4.13190372 5.67398324   9.805887      0 administration
#> 155   1 3.96970761 5.83617935   9.805887      0 administration
#> 157   1 4.11842105 5.68746591   9.805887      0 administration
#> 47    0 3.21901640 6.58687056   9.805887      0 administration
#> 178   1 2.49916908 7.30671788   9.805887      0 administration
#> 30    0 3.28341153 6.52247543   9.805887      0 administration
#> 26    0 0.77675620 9.02913076   9.805887      0 administration
#> 166   1 4.53491142 5.27097554   9.805887      0 administration
#> 177   1 2.13178324 7.67410372   9.805887      0 administration
#> 124   1 2.42660557 7.37928139   9.805887      0 administration
#> 168   1 3.84919615 5.95669081   9.805887      0 administration
#> 53    0 5.49738291 4.30850404   9.805887      0 administration
#> 20    0 5.68025733 4.12562963   9.805887      0 administration
#> 214   1 1.69523547 8.11065148   9.805887      0 administration
#> 23    0 4.69746044 5.10842651   9.805887      0 administration
#> 203   1 3.33165588 6.47423108   9.805887      0 administration
#> 99    0 4.31755994 5.48832702   9.805887      0 administration
#> 111   0 4.41183720 5.39404976   9.805887      0 administration
#> 150   1 2.27891912 7.52696784   9.805887      0 administration
#> 205   1 3.30839253 6.49749443   9.805887      0 administration
#> 96    0 4.16939665 5.63649030   9.805887      0 administration
#> 8     0 3.25869750 6.54718946   9.805887      0 administration
#> 46    0 0.07447530 9.73141166   9.805887      0 administration
#> 133   1 4.24246142 5.56342554   9.805887      0 administration
#> 183   1 5.95555827 3.85032869   9.805887      0 administration
#> 51    0 0.84018771 8.96569925   9.805887      0 administration
#> 29    0 1.98333135 7.82255561   9.805887      0 administration
#> 94    0 4.87568551 4.93020145   9.805887      0 administration
#> 195   1 0.33486370 9.47102326   9.805887      0 administration
#> 211   1 4.76239459 5.04349237   9.805887      0 administration
#> 154   1 4.34761947 5.45826748   9.805887      0 administration
#> 229   1 3.85631155 5.94957541   9.805887      0 administration
#> 115   0 4.76438628 5.04150067   9.805887      0 administration
#> 145   1 1.14169938 8.66418757   9.805887      0 administration
#> 61    0 3.54355271 6.26233425   9.805887      0 administration
#> 159   1 1.99964654 7.80624042   9.805887      0 administration
#> 103   0 3.93278475 5.87310221   9.805887      0 administration
#> 44    0 0.62048586 9.18540109   9.805887      0 administration
#> 198   1 1.41339454 8.39249242   9.805887      0 administration
#> 191   1 5.08104670 4.72484026   9.805887      0 administration
#> 40    0 5.82834859 3.97753837   9.805887      0 administration
#> 25    0 3.51180245 6.29408451   9.805887      0 administration
#> 43    0 4.25341933 5.55246762   9.805887      0 administration
#> 176   1 4.21153014 5.59435682   9.805887      0 administration
#> 39    0 3.33660229 6.46928467   9.805887      0 administration
#> 21    0 1.86452371 7.94136325   9.805887      0 administration
#> 52    0 2.63297393 7.17291302   9.805887      0 administration
#> 80    0 2.11090250 7.69498446   9.805887      0 administration
#> 116   0 4.48278259 5.32310437   9.805887      0 administration
#> 75    0 5.39689660 4.40899036   9.805887      0 administration
#> 137   1 4.11975637 5.68613058   9.805887      0 administration
#> 208   1 0.18642290 9.61946406   9.805887      0 administration
#> 197   1 1.27817250 8.52771446   9.805887      0 administration
#> 88    0 5.02353166 4.78235530   9.805887      0 administration
#> 200   1 0.42251828 9.38336867   9.805887      0 administration
#> 91    0 0.54477446 9.26111250   9.805887      0 administration
#> 32    0 4.39043603 5.41545093   9.805887      0 administration
#> 138   1 4.89939333 4.90649363   9.805887      0 administration
#> 147   1 2.94671942 6.85916753   9.805887      0 administration
#> 85    0 2.60019837 7.20568859   9.805887      0 administration
#> 5     0 4.32797080 5.47791615   9.805887      0 administration
#> 123   1 1.24162827 8.56425868   9.805887      0 administration
#> 33    0 2.21922027 7.58666669   9.805887      0 administration
#> 206   1 0.06354070 9.74234626   9.805887      0 administration
#> 223   1 1.31008018 8.49580678   9.805887      0 administration
#> 227   1 4.71419228 5.09169468   9.805887      0 administration
#> 60    0 1.36889160 8.43699536   9.805887      0 administration
#> 217   1 5.47672206 4.32916490   9.805887      0 administration
#> 105   0 2.16222254 7.64366441   9.805887      0 administration
#> 119   0 1.50638324 8.29950372   9.805887      0 administration
#> 126   1 5.32282243 4.48306453   9.805887      0 administration
#> 128   1 1.34565678 8.46023018   9.805887      0 administration
#> 226   1 2.07632016 7.72956679   9.805887      0 administration
#> 120   0 3.53986793 6.26601903   9.805887      0 administration
#> 187   1 0.53335641 9.27253054   9.805887      0 administration
#> 161   1 3.82253513 5.98335183   9.805887      0 administration
#> 59    0 3.42208700 6.38379996   9.805887      0 administration
#> 16    0 3.59398334 6.21190362   9.805887      0 administration
#> 95    0 3.97421127 5.83167569   9.805887      0 administration
#> 12    0 1.49325410 8.31263286   9.805887      0 administration
#> 142   1 1.88787520 7.91801175   9.805887      0 administration
#> 231   1 4.83402983 4.97185713   9.805887      0 administration
#> 92    0 1.35349009 8.45239687   9.805887      0 administration
#> 17    0 4.53871399 5.26717296   9.805887      0 administration
#> 202   1 0.29652404 9.50936292   9.805887      0 administration
#> 70    0 3.00468177 6.80120519   9.805887      0 administration
#> 84    0 5.68389547 4.12199149   9.805887      0 administration
#> 201   1 2.47030533 7.33558162   9.805887      0 administration
#> 117   0 5.16436960 4.64151735   9.805887      0 administration
#> 109   0 4.44753250 5.35835446   9.805887      0 administration
#> 238   1 3.36800003 6.43788693   9.805887      0 administration
#> 130   1 4.11473451 5.69115245   9.805887      0 administration
#> 230   1 3.80205567 6.00383128   9.805887      0 administration
#> 190   1 3.79348037 6.01240659   9.805887      0 administration
#> 28    0 0.59590128 9.20998568   9.805887      0 administration
#> 185   1 2.66646192 7.13942503   9.805887      0 administration
#> 35    0 2.45620011 7.34968685   9.805887      0 administration
#> 179   1 3.78529958 6.02058737   9.805887      0 administration
#> 66    0 2.31371248 7.49217447   9.805887      0 administration
#> 58    0 0.75299165 9.05289531   9.805887      0 administration
#> 78    0 3.83351865 5.97236831   9.805887      0 administration
#> 121   1 4.90810800 4.89777896   9.805887      0 administration
#> 2     0 0.24838638 9.55750057   9.805887      0 administration
#> 27    0 4.13037136 5.67551560   9.805887      0 administration
#> 216   1 3.59951870 6.20636826   9.805887      0 administration
#> 36    0 2.94353738 6.86234957   9.805887      0 administration
#> 186   1 3.39649643 6.40939052   9.805887      0 administration
#> 34    0 2.88884781 6.91703914   9.805887      0 administration
#> 164   1 2.57979269 7.22609427   9.805887      0 administration
#> 181   1 2.41991471 7.38597225   9.805887      0 administration
#> 232   1 5.68417358 4.12171338   9.805887      0 administration
#> 196   1 5.45222631 4.35366065   9.805887      0 administration
#> 162   1 3.56130638 6.24458058   9.805887      0 administration
#> 24    0 2.07149856 7.73438840   9.805887      0 administration
#> 160   1 3.32863901 6.47724795   9.805887      0 administration
#> 131   1 5.87040964 3.93547732   9.805887      0 administration
#> 132   1 3.51833232 6.28755463   9.805887      0 administration
#> 152   1 4.18763597 5.61825098   9.805887      0 administration
#> 100   0 4.01707529 5.78881167   9.805887      0 administration
#> 98    0 5.59368013 4.21220683   9.805887      0 administration
#> 172   1 0.95075018 8.85513678   9.805887      0 administration
#> 19    0 2.59621859 7.20966836   9.805887      0 administration
#> 82    0 0.50121657 9.30467038   9.805887      0 administration
#> 108   0 4.83804670 4.96784026   9.805887      0 administration
#> 148   1 0.99699479 8.80889217   9.805887      0 administration
#> 167   1 0.21107157 9.59481539   9.805887      0 administration
#> 182   1 1.01436405 8.79152290   9.805887      0 administration
#> 107   0 1.88888871 7.91699825   9.805887      0 administration
#> 189   1 0.27543950 9.53044746   9.805887      0 administration
#> 153   1 0.01709917 9.78878779   9.805887      0 administration
#> 64    0 3.80443574 6.00145122   9.805887      0 administration
#> 221   1 3.15681122 6.64907574   9.805887      0 administration
#> 74    0 3.12927879 6.67660817   9.805887      0 administration
#> 204   1 1.80256968 8.00331728   9.805887      0 administration
#> 7     0 0.01798826 9.78789869   9.805887      0 administration
#> 68    0 4.55899384 5.24689312   9.805887      0 administration
#> 122   1 2.37742921 7.42845775   9.805887      0 administration
#> 18    0 2.84036196 6.96552500   9.805887      0 administration
#> 106   0 1.59133233 8.21455463   9.805887      0 administration
#> 104   0 3.22958018 6.57630678   9.805887      0 administration
#> 87    0 0.18491288 9.62097408   9.805887      0 administration
#> 140   1 3.14300226 6.66288470   9.805887      0 administration
#> 1     0 1.69688604 8.10900092   9.805887      0 administration
#> 170   1 4.46875509 5.33713187   9.805887      0 administration
#> 14    0 5.72723586 4.07865109   9.805887      0 administration
#> 212   1 4.52729882 5.27858814   9.805887      0 administration
#> 134   1 2.21788846 7.58799850   9.805887      0 administration
#> 143   1 4.30891696 5.49697000   9.805887      0 administration
#> 156   1 5.98610977 3.81977719   9.805887      0 administration
#> 110   0 3.90697443 5.89891253   9.805887      0 administration
#> 220   1 4.16640138 5.63948558   9.805887      0 administration
#> 71    0 5.07716228 4.72872467   9.805887      0 administration
#> 237   1 2.04880951 7.75707745   9.805887      0 administration
#> 210   1 1.39213104 8.41375592   9.805887      0 administration
#> 215   1 5.93265641 3.87323054   9.805887      0 administration
#> 158   1 5.05750348 4.74838348   9.805887      0 administration
#> 49    0 0.14421634 9.66167062   9.805887      0 administration
#> 192   1 3.09510352 6.71078344   9.805887      0 administration
#> 144   1 0.35680733 9.44907963   9.805887      0 administration
#> 10    0 5.42424117 4.38164579   9.805887      0 administration
#> 174   1 4.57950058 5.22638638   9.805887      0 administration
#> 149   1 4.20068736 5.60519960   9.805887      0 administration
#> 234   1 0.79049236 9.01539460   9.805887      0 administration
#> 225   1 1.18711073 8.61877623   9.805887      0 administration
#> 125   1 3.18074041 6.62514655   9.805887      0 administration